Include anonymous sessions in logrocket
