+++ src/ui/types/index.ts
@@ -2,6 +2,7 @@ export interface User {
   name: string;
   picture: string;
   id: string;
+  internal: boolean;
 }
 
 export interface Recording {
